+++ CHANGES.md
@@ -1,3 +1,10 @@
+
+
+* Requires OCaml >= 4.08, Cmdliner >= 1.1.0, Fmt >= 0.9.0
+  and js_of_ocaml-compiler >= 4.0.0
+* Depend on the `js_of_ocaml-compiler.runtime` library rather than 
+  `js_of_ocaml`.
+  
 v0.7.0 2019-08-09 Zagreb
 ------------------------